# src/interfaces/http/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My Products API",
    "version": "v1"
  },
  "servers": [
    {
      "description": "Local server",
      "url": "/api"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "operationId": "listProducts",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "description": "limit number of results",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "description": "offset results from",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "order",
            "description": "order by attribute",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "asc(price)",
                  "desc(price)",
                  "asc(name)",
                  "desc(name)"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all Products",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#components/schemas/Product"
                          }
                        },
                        "count": {
                          "type": "number"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProduct",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "description": "Product Data",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Product"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Product not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "showProduct",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of product to show",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return product with given id",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Product"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProduct",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of product to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Product new data",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Product"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteProduct",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of product to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "operationId": "listCategories",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "description": "limit number of results",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "description": "offset results from",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all Categories",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Category"
                          }
                        },
                        "count": {
                          "type": "number"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCategory",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "description": "Category Data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Category"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Category not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "showCategory",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of category to show",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return category with given id",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Category"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateCategory",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of category to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Category new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Category"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCategory",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of category to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Category deleted successfully"
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      }
    },
    "/offers": {
      "get": {
        "operationId": "listOffers",
        "tags": [
          "Offers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "description": "limit number of results",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "description": "offset results from",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "itemUUID",
            "description": "search for offers that include item UUID",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all Offers",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Offer"
                          }
                        },
                        "count": {
                          "type": "number"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createOffer",
        "tags": [
          "Offers"
        ],
        "requestBody": {
          "description": "Offer Data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOffer"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Offer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Offer"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Offer not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{id}": {
      "get": {
        "operationId": "showOffer",
        "tags": [
          "Offers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of offer to show",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return offer with given id",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Offer"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Offer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateOffer",
        "tags": [
          "Offers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of offer to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Offer new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOffer"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Offer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/defaultResponses/DefaultProperties"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Offer"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Offer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteOffer",
        "tags": [
          "Offers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of offer to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Offer deleted successfully"
          },
          "404": {
            "description": "Offer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/defaultResponses/DefaultError"
                }
              }
            }
          }
        }
      }
    }
  },
  "defaultResponses": {
    "DefaultProperties": {
      "properties": {
        "statusCode": {
          "type": "number"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "DefaultError": {
      "allOf": [
        {
          "$ref": "#/defaultResponses/DefaultProperties"
        },
        {
          "properties": {
            "error": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "ValidationError"
                },
                "code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "path": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "NewProduct": {
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "image": {
            "type": "string",
            "format": "binary"
          },
          "categories": {
            "type": "array",
            "description": "UUID's of categories",
            "items": {
              "type": "string"
            }
          }
        },
        "type": "object"
      },
      "NewCategory": {
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "NewOffer": {
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "category",
              "product",
              "all"
            ]
          },
          "itemUUID": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "discount": {
            "type": "number"
          }
        }
      },
      "Product": {
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Unique Identifier representing a product",
            "x-primary-key": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "Category": {
        "properties": {
          "uuid": {
            "type": "string",
            "x-primary-key": true
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "Offer": {
        "properties": {
          "uuid": {
            "type": "string",
            "x-primary-key": true
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "category",
              "product",
              "all"
            ]
          },
          "itemUUID": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "discount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      }
    }
  }
}